+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Add mutex to podready function [#158](https://github.com/nre-learning/antidote-core/pull/158)
 - Make k8s configuration configurable at Antidote level [#159](https://github.com/nre-learning/antidote-core/pull/159)
 - Changed resource order on ingest, and added early return on error [#160](https://github.com/nre-learning/antidote-core/pull/160)
+- Add "Authors" to lesson metadata [#162](https://github.com/nre-learning/antidote-core/pull/162)
 
 ## v0.5.1 - February 17, 2020